feat: add MBU metric, JIT decode benchmark, and teacher-forcing verification
- benchmark.py: add MBU (Memory Bandwidth Utilization) for decode phase,
  update TPU v6e peak to 918 TFLOPS, use JittedModel for decode benchmark
- verify_kvcache.py: replace exact-match generation test with teacher-forcing
  logit comparison (tolerant of bfloat16 precision), skip slow non-cached
  path by default (--full to enable)
- runner.py: fix hardcoded bfloat16 mask dtype in greedy_generate_no_cache

+++ runner.py
@@ -396,7 +396,7 @@ def greedy_generate_no_cache(
     for _ in range(max_new_tokens):
         seq_len = current_ids.shape[1]
         positions = make_positions(seq_len, current_ids.shape[0])
-        mask = make_causal_mask(seq_len, dtype=jnp.bfloat16)
+        mask = make_causal_mask(seq_len, dtype=model.dtype)
 
         logits, _ = model(current_ids, positions, mask)
         next_token = jnp.argmax(logits[:, -1, :], axis=-1, keepdims=True)
